requirements.txt
Flask==2.0.3
python-telegram-bot==20.3
gunicorn
aiolimiter==1.1.0
apscheduler==3.6.3
requests==2.32.3
SQLAlchemy==2.0.36
